start of game

+++ include/game_card.h
@@ -39,7 +39,7 @@ struct Card {
     std::vector<std::vector<Tile>> mTiles;
 
     sf::Sprite mSprite;
-    sf::Vector2f mPosition;
+    sf::Vector2i mPosition;
 };
 
 }  // namespace game_model
